Added support for the "run" CLI task runner
The format of the justfile and the Runfile defaults for the two task runners is bery close.  The main difference is in how they handle variables, sub-shells, and parameters to their tasks.

"just" calls tasks "recipes" and "run" calss tasks "commands" so I am just going to call them "tasks."

This new feature is intended for version 1.2.4

This first commit is only for the Ruby gem implementation.