clean_up README files and codebase
The branch "support_run" has been archived.  Some of the codebase improvements have been moved from that branch to this one.  Hopefully the new stuff that I put inplace to support the run task runner will be usefully if any other task runner pops up on the radar.

The make README in the repo run has been cut down to a nice easily understood size.  The intent is to have all of the documentation in the github wiki for the project.