no_brag Added the --no-brag CLI option
Adding command-line options to control the business logic required that constants become class variables.  That meant that all of the global methods had to become instance methods.

The --no-brag option tells justprep to no put the header/footer wrapper around the included content.

As of this commit there are no tests for this new option.